requirements.txt
netifaces==0.10.4
pyzmq==22.3.0
requests==2.22.0